+++ build.gradle
@@ -32,8 +32,7 @@ dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-validation'
 	implementation 'org.springframework.boot:spring-boot-starter-oauth2-client'
 	implementation 'org.thymeleaf.extras:thymeleaf-extras-springsecurity5'
-
-
+	implementation 'org.mariadb.jdbc:mariadb-java-client'
 
 
 	asciidoctorExtensions 'org.springframework.restdocs:spring-restdocs-asciidoctor'
@@ -46,6 +45,9 @@ dependencies {
 	testImplementation 'org.springframework.restdocs:spring-restdocs-mockmvc'
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 	testImplementation 'org.springframework.security:spring-security-test'
+
+
+
 }
 
 test {
